automation/requirements.txt
crewai>=0.70.0
crewai-tools>=0.12.0
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
